Deprecate usage of XML Spring schemas
Move Settings13 test

Related to #141.
